refactor: add ownership check in TaskServiceImpl using username validation

+++ src/main/java/com/pen/taskmanagement/service/TaskServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import com.pen.taskmanagement.dtos.TaskRequest;
 import com.pen.taskmanagement.dtos.TaskResponse;
+import com.pen.taskmanagement.exceptions.ForbiddenException;
 import com.pen.taskmanagement.exceptions.ResourceNotFoundException;
 import com.pen.taskmanagement.mapper.TaskMapper;
 import com.pen.taskmanagement.model.Project;
@@ -15,6 +16,7 @@ import com.pen.taskmanagement.model.User;
 import com.pen.taskmanagement.repository.ProjectRepository;
 import com.pen.taskmanagement.repository.TaskRepository;
 import com.pen.taskmanagement.repository.UserRepository;
+import com.pen.taskmanagement.utilities.SecurityUtil;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -28,6 +30,7 @@ public class TaskServiceImpl implements TaskService {
     private final UserRepository userRepository;
     private final ProjectRepository projectRepository;
     private final TaskMapper taskMapper;
+    private final SecurityUtil securityUtil;
 
     @Override
     public TaskResponse createTask(TaskRequest taskRequest) {
@@ -42,7 +45,6 @@ public class TaskServiceImpl implements TaskService {
         Task task = taskMapper.toEntity(taskRequest, user, project);
         task.setStartTime(LocalDateTime.now());
         TaskResponse taskResponse = taskMapper.toResponse(taskRepository.save(task));
-
 
         return taskResponse;
     }
@@ -69,7 +71,15 @@ public class TaskServiceImpl implements TaskService {
         if (!taskRepository.existsById(id))
             throw new ResourceNotFoundException("Task not found");
 
-        taskRepository.deleteById(id);
+        Task task = taskRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Task not found"));
+
+        String username = securityUtil.extractUsername();
+
+        if (username.equals(task.getUser().getUsername())) {
+            taskRepository.deleteById(id);
+        } else {
+            throw new ForbiddenException("Permissions not found");
+        }
 
     }
 
@@ -81,12 +91,18 @@ public class TaskServiceImpl implements TaskService {
         Project project = projectRepository.findById(taskRequest.projectId())
                 .orElseThrow(() -> new ResourceNotFoundException("Project not found"));
         Task task = taskRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Task not found"));
-        task.setName(taskRequest.name());
-        task.setDescription(taskRequest.description());
-        // task.setStartTime(taskRequest.startDateTime());
-        task.setEndTime(taskRequest.endDateTime());
-        task.setProject(project);
-        task.setUser(user);
+        String username = securityUtil.extractUsername();
+
+        if (username.equals(task.getUser().getUsername())) {
+            task.setName(taskRequest.name());
+            task.setDescription(taskRequest.description());
+            // task.setStartTime(taskRequest.startDateTime());
+            task.setEndTime(taskRequest.endDateTime());
+            task.setProject(project);
+            task.setUser(user);
+        } else {
+            throw new ForbiddenException("Permissions not found");
+        }
 
         return taskMapper.toResponse(taskRepository.save(task));
     }
